# scripts/openapi3-spec.yml
openapi: 3.0.0
info:
  title: 'Gụọ API'
  description: Gụọ API enables users query data on genres, series, books and their authors through an intuitive API schema as defined by the [OpenAPI Specification v3](https://swagger.io/specification/). It virtually simulate the structure of a library
  termsOfService: https://github.com/Adophilus/guo-api/blob/main/TERMS-OF-SERVICE.txt
  contact:
    email: uchenna19of@gmail.com
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  version: '1.0.0'
...
servers:
  - url: 'https://guo-api.adophilus.xyz/api'
paths:
...
  /series:
    post:
      tags:
      - series
      summary: Add a new series to the library
      description: Using this endpoint, a new series collection can be added to the
        library
      operationId: addSeries
      requestBody:
        description: The series that needs to be added to the library
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SeriesPost'
        required: true
      responses:
        '201':
          description: Series created successfully!
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Series'
        '400':
          description: Invalid Series ID!
        '409':
          description: Series ID taken!
    get:
      summary: 'Fetch all series from the library'
      description: Using this endpoint, all series collections can be fetched from the library
      operationId: getSeries
      tags:
        - series
      responses:
        '200':
          description: 'Returns a list of series'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SeriesSlim'
  /series/{seriesId}:
    get:
      tags:
      - series
      summary: Fetch series by {seriesId} from the library
      description: Using this endpoint, the series with a corresponding {seriesId} will be fetched from the library
      operationId: getSeriesById
      parameters:
      - name: seriesId
        in: path
        required: true
        schema:
          type: string
      responses:
        200:
          description: Returns an Series object
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Series'
        404:
          description: The series with a matching {seriesID} was not found in the library
    put:
      tags:
      - series
      summary: Update an existing series with a matching {seriesId}
      description: Using this endpoint, an existing series with a matching {seriesId} in the library can be modified
      operationId: updateSeries
      parameters:
      - name: seriesId
        in: path
        required: true
        schema:
          type: string
      requestBody:
        description: The series data that needs to be modified in the library
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SeriesPut'
        required: true
      responses:
        200:
          description: Returns an Series object
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Series'
        404:
          description: The series with a matching {seriesID} was not found in the library
        400:
          description: Invalid Series data!
        409:
          description: Series ID taken!
    delete:
      summary: 'Delete an existing Series with a matching {SeriesId}'
      description: Using this endpoint, an existing Series collections with a matching {SeriesId} can be deleted from the library
      operationId: deleteSeries
      parameters:
      - name: seriesId
        in: path
        required: true
        schema:
          type: string
      tags:
        - Series
      responses:
        '410':
          description: 'The Series with a matching {SeriesId} has been deleted!'
...
components:
  schemas:
...
    Series:
      type: "object"
      properties:
        id:
          type: "string"
          example: "SC-101"
        name:
          type: "string"
          example: "Amazing series"
        plannedVolumes:
          type: "integer"
          description: The number of volumes the series was intended to have
          example: 1
          default: 1
        bookTourEvents:
          type: "integer"
          description: The number of book tour events the series has/had
          example: 3
          default: 0
        books:
          type: "array"
          items:
            $ref: '#/components/schemas/BookSlim'
    SeriesPost:
      type: object
      required:
      - id
      - name
      - plannedVolumes
      - bookTourEvents
      properties:
        id:
          type: "string"
          example: "SC-101"
        name:
          type: "string"
          example: "Amazing series"
        plannedVolumes:
          type: "integer"
          description: The number of volumes the series was intended to have
          example: 1
          default: 1
        bookTourEvents:
          type: "integer"
          description: The number of book tour events the series has/had
          example: 3
          default: 0
        books:
          type: "array"
          description: A list book IDs belonging to this genre
          items:
            type: string
            description: book ID
    SeriesPut:
      type: object
      properties:
        id:
          type: "string"
          example: "SC-101"
        name:
          type: "string"
          example: "Amazing series"
        plannedVolumes:
          type: "integer"
          description: The number of volumes the series was intended to have
          example: 1
          default: 1
        bookTourEvents:
          type: "integer"
          description: The number of book tour events the series has/had
          example: 3
          default: 0
        books:
          type: "array"
          description: A list book IDs belonging to this genre
          items:
            type: string
            description: book ID
    SeriesSlim:
      type: object
      properties:
        id:
          type: "string"
          example: "SC-101"
        name:
          type: "string"
          example: "Amazing series"
        plannedVolumes:
          type: "integer"
          description: The number of volumes the series was intended to have
          example: 1
          default: 1
        bookTourEvents:
          type: "integer"
          description: The number of book tour events the series has/had
          example: 3
          default: 0
        books:
          type: "array"
          description: A list book IDs belonging to this genre
          items:
            type: string
            description: book ID
...
    BookSlim:
      type: object
      properties:
        id:
          type: "string"
          example: "BK101"
        title:
          type: "string"
          example: "Amazing Book"
        volume:
          type: "integer"
          example: 1
        comment:
          type: "string"
          example: "This book is the most amazing book you'll ever read"
          description: The comment of the book
        authors:
          type: "array"
          description: author ID
          example: "[ '{authorsId1}', ... ]"
          items:
            type: "string"
        genres:
          type: "array"
          description: genre ID
          example: "[ {genreId1}, ... ]"
          items:
            type: "integer"
        series:
          type: "array"
          description: series ID
          example: "[ '{seriesId1}', ... ]"
          items:
            type: "string"